+++ src/ui/components/organisms/RepositoryBrowser/Directory.tsx
@@ -1,22 +1,23 @@
-import range from "lodash/range";
-import path from "path";
-import React from "react";
-import { ContextMenuProvider } from "react-contexify";
-import FaFile from "react-icons/fa/file";
-import FaFolder from "react-icons/fa/folder";
-import FaFolderOpen from "react-icons/fa/folder-open";
-import FaTrash from "react-icons/fa/trash";
-import { connect } from "react-redux";
-import Tooltip from "react-tooltip";
-import { lifecycle } from "recompose";
-import styled from "styled-components";
-import { readFileStats } from "../../../../domain/filesystem/queries/readFileStats";
-import { FileInfo } from "../../../../domain/types";
-import { RootState } from "../../../reducers";
-import * as RepositoryActions from "../../../reducers/repository";
-import { AddDir } from "./AddDir";
-import { AddFile } from "./AddFile";
-import { File } from "./File";
+import range from "lodash/range"
+import path from "path"
+import React from "react"
+import { ContextMenuProvider } from "react-contexify"
+import FaFile from "react-icons/fa/file"
+import FaFolder from "react-icons/fa/folder"
+import FaFolderOpen from "react-icons/fa/folder-open"
+import FaTrash from "react-icons/fa/trash"
+import { connect } from "react-redux"
+import Tooltip from "react-tooltip"
+import { lifecycle } from "recompose"
+import styled from "styled-components"
+import { readFileStats } from "../../../../domain/filesystem/queries/readFileStats"
+import { FileInfo } from "../../../../domain/types"
+import * as EditorActions from "../../../actions/editorActions"
+import { RootState } from "../../../reducers"
+import * as RepositoryActions from "../../../reducers/repository"
+import { AddDir } from "./AddDir"
+import { AddFile } from "./AddFile"
+import { File } from "./File"
 
 type OwnProps = {
   root: string
@@ -29,7 +30,7 @@ type OwnProps = {
 const actions = {
   startFileCreating: RepositoryActions.startFileCreating,
   startDirCreating: RepositoryActions.startDirCreating,
-  deleteDirectory: RepositoryActions.deleteDirectory
+  deleteDirectory: EditorActions.deleteDirectory
 }
 
 type Props = OwnProps &
@@ -53,7 +54,7 @@ export const Directory: React.ComponentType<OwnProps> = connect(
   (state: RootState, ownProps: OwnProps) => {
     return {
       ...ownProps,
-      editingFilepath: state.editor.filepath,
+      editingFilepath: state.buffer.filepath,
       touchCounter: state.repository.touchCounter,
       isFileCreating: ownProps.dirpath === state.repository.fileCreatingDir,
       isDirCreating: ownProps.dirpath === state.repository.dirCreatingDir
@@ -143,7 +144,9 @@ export const Directory: React.ComponentType<OwnProps> = connect(
                         onClick={ev => {
                           ev.stopPropagation()
                           this.setState({ opened: true }, () => {
-                            this.props.startFileCreating(dirpath)
+                            this.props.startFileCreating({
+                              fileCreatingDir: dirpath
+                            })
                           })
                         }}
                       />
@@ -153,7 +156,9 @@ export const Directory: React.ComponentType<OwnProps> = connect(
                         onClick={ev => {
                           ev.stopPropagation()
                           this.setState({ opened: true }, () => {
-                            this.props.startDirCreating(dirpath)
+                            this.props.startDirCreating({
+                              dirCreatingDir: dirpath
+                            })
                           })
                         }}
                       />
@@ -165,7 +170,7 @@ export const Directory: React.ComponentType<OwnProps> = connect(
                             onClick={ev => {
                               ev.stopPropagation()
                               if (confirm(`Confirm: delete ${dirpath}`)) {
-                                this.props.deleteDirectory(dirpath)
+                                this.props.deleteDirectory({ dirpath })
                               }
                             }}
                           />